Extract AmazonS3 object creation to separate module
